Base.Cartesian

Base.Cartesian

El módulo Cartesian (no exportado) proporciona macros que facilitan escribir algoritmos multidimensionales. Es deseable que, a largo plazo, este módulo Cartesian no sea necesario; sin embargo, en la actualidad es una de las pocas formas de escribir código multidimensional compacto y con rendimiento.

Principios de uso

Un ejemplo de uso simple podría ser:

@nloops 3 i A begin
    s += @nref 3 A i
end

que genera el siguiente código:

for i_3 = 1:size(A,3)
    for i_2 = 1:size(A,2)
        for i_1 = 1:size(A,1)
            s += A[i_1,i_2,i_3]
        end
    end
end

En general, Cartesian permitirá escribir código que contiene elementos repetitivos, como los bucles anidados de este ejemplo. Otras aplicaciones incluyen expresiones repetidas (por ejemplo, desenrollado de bucles) o crear llamadas a función con números variables de argumentos sin usar la construcción "splat" (i...).

Sintaxis Básica

La sintaxis básica de @nloops es la siguiente:

* El primer argumento debe ser un entero (no una variable) que especifica el número de bucles. * El segundo argumento es el prefijo simbólico que se utilizará para la variable iteradora. De este modo, en el ejemplo anterior usamos i, y se generaron las variables i_1, i_2, i_3. * El tercer argumento especifica el rango para cada variable iteradora. Si se usa una variable (símbolo) aquí, es considerado como 1:size(A,dim). De forma más flexible, se puede usar la sintaxis de expresiones basadas en funciones anónimas que se decribe más adelante. * El último argumento es el cuerpo del bucle. En el ejemplo anterior, lo que aparece entre begin...end.

Hay otras características adicionales de @nloops descritas en la sección de referencia.

@nref sigue un patrn similar, generando A[i_1,i_2,i_3] a partir de @nref 3 A i. La práctica general es leer de izquierda a derecha, por lo que @nloops es @nloops 3 i A expr (como en el bucle for i_2 = 1:size(A,2), donde i_2 está a la izquierda y el rango a la derecha) mientras que @nref es @nref 3 A i (como en A[i_1,i_2,i_3], donde el array va primero).

Si ests desarrollando código con Cartesian, puedes encontrar que depurar es más sencillo cuando examinas el código generado, usando macroexpand:

julia> macroexpand(:(@nref 2 A i))
:(A[i_1, i_2])

Proporcionando el número de expresiones

El primer argumentos de estas dos macros es el número de expresiones, que debe ser un entero. Cuando estás escribiendo una funcin que pretendes que trabaje en múltiples dimensiones, esto puede no ser algo que desees codificar. Si estás escribiendo código que necesitas que trabaje con versiones antiguas de Julia, deberías usar la macro @ngenerate descrita en una versión más antigua de esta documentación.

Empezando en Julia 0.4-pre, el enfoque recomendado es usar una @generated function. He aquí un ejemplo:

@generated function mysum(A::Array{T,N}) where {T,N}
    quote
        s = zero(T)
        @nloops $N i A begin
            s += @nref $N A i
        end
        s
    end
end

Naturalmente, también podemos preparar expresiones o realizar cálculos antes del bloque quote.

Expresiones función anónima como argumentos de macros

Quizás la característica más potente de Cartesian es la capacidad de proporcionar expresiones función-anónima que son evaluadas en tiempo de análisis sintáctico. Consideremos un ejemplo sencillo:

@nexprs 2 j->(i_j = 1)

@nexprs genera n expresiones que siguen un patrón. Este código generaría las siguientes instrucciones:

i_1 = 1
i_2 = 1

En cada instrucción generada un j aislado (la variable de la función anónima) es reemplazada por valores en el rango 1:2. Hablando de forma general, Cartesian emplea una sintaxis parecida a LaTeX. Esto te permite hacer operaciones sobre el índice j. He aquí un ejemplo que calcula los pasos de un array:

s_1 = 1
@nexprs 3 j->(s_{j+1} = s_j * size(A, j))

generará las expresiones

s_1 = 1
s_2 = s_1 * size(A, 1)
s_3 = s_2 * size(A, 2)
s_4 = s_3 * size(A, 3)

Las expresiones función anónima tienen muchos usos en la práctica.

Referencia de las Macros

Base.Cartesian.@nloops — Macro.

@nloops N itersym rangeexpr bodyexpr
@nloops N itersym rangeexpr preexpr bodyexpr
@nloops N itersym rangeexpr preexpr postexpr bodyexpr

Generate N nested loops, using itersym as the prefix for the iteration variables. rangeexpr may be an anonymous-function expression, or a simple symbol var in which case the range is indices(var, d) for dimension d.

Optionally, you can provide "pre" and "post" expressions. These get executed first and last, respectively, in the body of each loop. For example:

@nloops 2 i A d -> j_d = min(i_d, 5) begin
    s += @nref 2 A j
end

would generate:

for i_2 = indices(A, 2)
    j_2 = min(i_2, 5)
    for i_1 = indices(A, 1)
        j_1 = min(i_1, 5)
        s += A[j_1, j_2]
    end
end

If you want just a post-expression, supply nothing for the pre-expression. Using parentheses and semicolons, you can supply multi-statement expressions.

source

Base.Cartesian.@nref — Macro.

@nref N A indexexpr

Generate expressions like A[i_1, i_2, ...]. indexexpr can either be an iteration-symbol prefix, or an anonymous-function expression.

julia> @macroexpand Base.Cartesian.@nref 3 A i
:(A[i_1, i_2, i_3])

source

Base.Cartesian.@nextract — Macro.

@nextract N esym isym

Generate N variables esym_1, esym_2, ..., esym_N to extract values from isym. isym can be either a Symbol or anonymous-function expression.

@nextract 2 x y would generate

x_1 = y[1]
x_2 = y[2]

while @nextract 3 x d->y[2d-1] yields

x_1 = y[1]
x_2 = y[3]
x_3 = y[5]

source

Base.Cartesian.@nexprs — Macro.

@nexprs N expr

Generate N expressions. expr should be an anonymous-function expression.

julia> @macroexpand Base.Cartesian.@nexprs 4 i -> y[i] = A[i+j]
quote
    y[1] = A[1 + j]
    y[2] = A[2 + j]
    y[3] = A[3 + j]
    y[4] = A[4 + j]
end

source

Base.Cartesian.@ncall — Macro.

@ncall N f sym...

Generate a function call expression. sym represents any number of function arguments, the last of which may be an anonymous-function expression and is expanded into N arguments.

For example @ncall 3 func a generates

func(a_1, a_2, a_3)

while @ncall 2 func a b i->c[i] yields

func(a, b, c[1], c[2])

source

Base.Cartesian.@ntuple — Macro.

@ntuple N expr

Generates an N-tuple. @ntuple 2 i would generate (i_1, i_2), and @ntuple 2 k->k+1 would generate (2,3).

source

Base.Cartesian.@nall — Macro.

@nall N expr

Check whether all of the expressions generated by the anonymous-function expression expr evaluate to true.

@nall 3 d->(i_d > 1) would generate the expression (i_1 > 1 && i_2 > 1 && i_3 > 1). This can be convenient for bounds-checking.

source

Base.Cartesian.@nany — Macro.

@nany N expr

Check whether any of the expressions generated by the anonymous-function expression expr evaluate to true.

@nany 3 d->(i_d > 1) would generate the expression (i_1 > 1 || i_2 > 1 || i_3 > 1).

source

Base.Cartesian.@nif — Macro.

@nif N conditionexpr expr
@nif N conditionexpr expr elseexpr

Generates a sequence of if ... elseif ... else ... end statements. For example:

@nif 3 d->(i_d >= size(A,d)) d->(error("Dimension ", d, " too big")) d->println("All OK")

would generate:

if i_1 > size(A, 1)
    error("Dimension ", 1, " too big")
elseif i_2 > size(A, 2)
    error("Dimension ", 2, " too big")
else
    println("All OK")
end

source